include bash
minor adjustment to ignore symlinked files

+++ script/mkrestore.sh
@@ -14,26 +14,25 @@ DOTFILES=(
     [neofetch]="$CONFIG_DIR/neofetch"
     [Kvantum]="$CONFIG_DIR/Kvantum"
     [konsole]="$LOCAL_DIR/share/konsole"
+    [bash]="$HOME/.bashrc"
 )
 
 for key in "${!DOTFILES[@]}"; do
     TARGET="${DOTFILES[$key]}"
     BACKUP="$BACKUP_DIR/$(basename "$TARGET")"
 
-    if [ -L "$TARGET" ]; then
-        echo "Removing symwink for $key.."
+    if [ -e "$BACKUP" ] && [ -L "$TARGET" ]; then
+        echo "Cleaning symlink for $key.."
         rm "$TARGET"
     else
-        echo "$key is not symwinked, skipping.."
+        echo "$key backup doesn't exist, skipping.."
         continue
     fi
 
     if [ -e "$BACKUP" ]; then
         echo "Restoring your $key from backup"
         mv "$BACKUP" "$TARGET"
-    else 
-        echo "No backup found $key, silly!"
     fi
 done
 
-echo "Restore compweted! You might want to reboot to start over~"
+echo "Restore completed! You might want to logout to refresh~"
